move futures to eg_futures
